pyproject.toml
[project]
name = "app" # startProject 로 설정한 이름을 넣어줌.
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.13"


[tool.poetry.dependencies] # 기본 환경
django = ">=5.1.7,<6.0.0"
djangorestframework = ">=3.16.0,<4.0.0"
python-decouple = "^3.8"
psycopg = "^3.2.6"
python-dotenv = "^1.1.0"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
requests = "^2.32.3"
django-filter = "^25.1"
django-storages = "^1.14.6"
boto3 = "^1.37.28"
certifi = "^2025.1.31"

[tool.poetry.group.prod.dependencies] # 배포 환경
isort = "^6.0.1"    # import를 알파벳 순으로 자동 정렬
black = "^25.1.0"   # 코드 포매팅 (PEP8 기준)

[tool.poetry.group.dev.dependencies] # 개발 환경
drf-spectacular = "^0.28.0" # DRF용 API 문서 자동 생성 (OPENAPI 3.0 기반)
ipython = "^9.0.2"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
combine_as_imports = true  # ← 이건 추가해도 OK

[tool.black]
line-length = 88
